Created vcastCommandRunner to contain all of the code that invokes clicast or vpython
